requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
bidict==0.22.1
channels==4.0.0
channels-redis==4.1.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
msgpack==1.0.5
PyJWT==2.6.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
redis==4.5.4
sqlparse==0.4.3
tzdata==2023.3
